pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "dvc[gdrive]>=3.66.0",
    "gdown>=5.2.0",
    "hydra-core==1.3.2",
    "invoke>=2.2.1",
    "ipykernel>=7.1.0",
    "matplotlib==3.10.8",
    "mnist==0.2.2",
    "numpy==2.4.0",
    "omegaconf==2.3.0",
    "pre-commit>=4.5.1",
    "pytorch-lightning==2.6.0",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "torch==2.9.1",
    "torch-ema==0.3",
    "types-requests>=2.32.4.20250913",
    "wandb==0.23.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.mypy]
disable_error_code = ["override", "assignment"]
ignore_missing_imports = true

[tool.ruff]
line-length=120
